move plugin class to root directory
